requirements.txt
requests
PyYAML
pandas
ipykernel
psycopg2
numpy
sqlparse
streamlit
scikit-learn